backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
bcrypt==4.1.2
PyJWT==2.8.0
googletrans==4.0.0rc1
gTTS==2.5.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
werkzeug==3.0.1
